Use member selection for virtual interface vars

+++ src/V3Dead.cpp
@@ -321,7 +321,7 @@ private:
     }
     bool mightElimVar(AstVar* nodep) const {
         if (nodep->isSigPublic()) return false;  // Can't elim publics!
-        if (nodep->isIO() || nodep->isClassMember()) return false;
+        if (nodep->isIO() || nodep->isClassMember() || nodep->isUsedVirtIface()) return false;
         if (nodep->isTemp() && !nodep->isTrace()) return true;
         return m_elimUserVars;  // Post-Trace can kill most anything
     }
